syz-manager: fix uint size of coverage bitmap size in covfilter test
createCoverageBitmap() returns a 64bit uint value as the coverage bitmap
size.

$ go test -v -count=1 ./syz-manager

+++ syz-manager/covfilter_test.go
@@ -24,7 +24,7 @@ func TestCreateBitmap(t *testing.T) {
 	}
 	bitmap := createCoverageBitmap(cfg, pcs)
 	start := order.Uint64(bitmap[0:])
-	size := order.Uint32(bitmap[8:])
+	size := order.Uint64(bitmap[8:])
 	if start != 0x81000002 || size != 0x20001b {
 		t.Fatalf("bad region 0x%x/0x%x", start, size)
 	}
